pyproject.toml
[tool.poetry]
name = "dvmn-notification-bot"
version = "0.1.0"
description = ""
authors = ["Roman Okatev <43115066+Nebulawalker@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "dvmn_notification_bot"}]

[tool.poetry.dependencies]
python = "^3.11"
environs = "^9.5.0"
requests = "^2.30.0"
python-telegram-bot = "13.15"
aiogram = "^2.25.1"
httpx = "^0.24.0"
loguru = "^0.7.0"
notifiers = "^1.3.3"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
